requirements.txt
Flask==3.0.0
Flask-Cors==4.0.1
gunicorn==22.0.0
PyPDF2==3.0.1
requests==2.31.0
google-generativeai==0.4.0
python-dotenv==1.0.1